+++ genomenotebook/__init__.py
@@ -1,7 +1,10 @@
-__version__ = "0.5.1"
+__version__ = "0.5.2"
 
 from .data import get_example_data_dir
 from .browser import GenomeBrowser
-from .utils import get_genome_annotations, get_genes_from_annotation
+from .utils import (get_genome_annotations, 
+                    get_features_from_annotation, 
+                    default_glyphs
+                   )
 from bokeh.io import output_notebook
 output_notebook(hide_banner=True)